backend/rag_pipeline/requirements-complete.txt
# RAG Pipeline 完整套件清單
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# 資料處理套件
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# 向量搜尋套件
pymilvus>=2.3.0

# NLP 套件
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0

# CrewAI 和 LangChain 套件
crewai>=0.1.0
langchain>=0.0.300
langchain-community>=0.0.10
langchain-openai>=0.0.1
langchain-anthropic>=0.0.1
langchain-google-genai>=0.0.1

# AI 模型套件
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# 資料庫套件
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# HTTP 客戶端套件
httpx>=0.24.0
aiohttp>=3.8.0
requests>=2.31.0

# 工具套件
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0.0

# 日誌與監控套件
structlog>=23.1.0
prometheus-client>=0.17.0

# 工具函數套件
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# 資料處理套件
openpyxl>=3.1.0
xlrd>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 效能優化套件
orjson>=3.9.0
ujson>=5.8.0
uvloop>=0.17.0

# 其他實用套件
tiktoken>=0.5.0
tqdm>=4.65.0
click>=8.1.0
rich>=13.0.0
typer>=0.9.0
toml>=0.10.0

# 開發工具
ipython>=8.0.0
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0 